Move Calculator logic to utils
