+++ plasmapy/utils/calculator/favicon.ico
