+++ plasmapy/utils/calculator/main_interface.py
@@ -1,9 +1,14 @@
+"""
+Collection of private functions to load properties and construct widgets
+"""
+
 import astropy.units as units
 import json
+import os
 
 from ipywidgets import widgets
 
-from .widget_helpers import (
+from plasmapy.utils.calculator.widget_helpers import (
     _calculate_button,
     _CheckBox,
     _clear_button,
@@ -19,6 +24,11 @@ from .widget_helpers import (
 light_blue = "#00BFD8"
 light_gray = "#A9A9A9"
 
+__all__ = ["_create_interactive_layout", "_create_output_layout", "test"]
+
+"""
+Grid Data specifies how the input section is layed out
+"""
 grid_data = [
     [
         _create_label("Parameter", color=light_blue),
@@ -110,9 +120,10 @@ grid_data = [
     ],
 ]
 
-
 def _create_interactive_layout():
-    ## grid config
+    """
+    Interactive grid layout for input parameters populated in grid_data
+    """
     grid = widgets.GridspecLayout(18, 3)
     grid.layout.margin = "10px"
 
@@ -126,10 +137,13 @@ def _create_interactive_layout():
 
 
 def _create_output_layout():
+    """
+    Tab layout for output parameters, populated from properties_metadata.json
+    """
     app = widgets.Tab()
     children = []
 
-    with open("plasma_calculator_helper/properties_metadata.json") as f:
+    with open("properties_metadata.json") as f:
         data = json.load(f)
 
     for i, title in enumerate(data):
