+++ plasmapy/utils/calculator/properties_metadata.json
